Renames back to eventhub as that's more prevalent
